# app/Http/Controllers/Gig/GigquestionController.php
<?php

namespace App\Http\Controllers\Gig;

use App\Http\Controllers\Controller;
use App\Http\Requests\Gigquestion;
use Illuminate\Http\Request;
use App\Repository\GigquestionRepositoryInterface;

class GigquestionController extends Controller
{
    private $gigquestionInterface;

    public function __construct(GigquestionRepositoryInterface $gigquestionInterface)
    {
        $this->gigquestionInterface = $gigquestionInterface;
    }
    /**
     * Display a listing of the resource.
     *
     * @return \Illuminate\Http\Response
     */

     /**
     * @OA\Get(
     ** path="/gigquestion/get/{gigId}",
     *   tags={"Gig Question"},
     *   summary="Get User Gig Questions",
     *   operationId="getusergigquestion",
     *   security={{"bearer_token":{}}},
     *   @OA\Parameter(
     *          name="gigId",
     *          required=true,
     *          in="path",
     *          @OA\Schema(
     *              type="integer"
     *          )
     *   ),
     *   @OA\Response(
     *      response=200,
     *       description="Success",
     *      @OA\MediaType(
     *           mediaType="application/json",
     *      )
     *   ),
     *   @OA\Response(
     *      response=400,
     *      description="Bad Request"
     *   ),
     *      @OA\Response(
     *          response=403,
     *          description="Forbidden"
     *      )
     *)
     **/
    public function index($id)
    {
        try{
            $gigquestions = $this->gigquestionInterface->collectionbycondition([ 'userId' => Auth()->user()->id, 'gigId' => $id ]);
            return response()->json(['success' => true, 'gigquestions' => $gigquestions ]);
        }catch(Exception $ex)
        {
            return response()->json(['success' => true, 'error' => array('message' => $ex->getMessage() ) ]);
        }
    }

    /**
     * Show the form for creating a new resource.
     *
     * @return \Illuminate\Http\Response
     */
    public function create()
    {
        //
    }

    /**
     * Store a newly created resource in storage.
     *
     * @param  \Illuminate\Http\Request  $request
     * @return \Illuminate\Http\Response
     */

     /**
     * @OA\Post(
     ** path="/gigquestion/store",
     *   tags={"Gig Question"},
     *   summary="Create New GiG Question",
     *   operationId="createnewquestion",
     *   security={{"bearer_token":{}}},
     *
     *   @OA\RequestBody(
     *         required=true,
     *         @OA\MediaType(
     *             mediaType="application/json",
     *             @OA\Schema(
     *                 @OA\Property(
     *                     property="question",
     *                     type="string"
     *                 ),
     *                 @OA\Property(
     *                     property="gigId",
     *                     type="integer"
     *                 )
     *             )
     *         )
     *     ),
     *   @OA\Response(
     *      response=201,
     *       description="Success",
     *      @OA\MediaType(
     *           mediaType="application/json",
     *      )
     *   ),
     *   @OA\Response(
     *      response=400,
     *      description="Bad Request"
     *   ),
     *      @OA\Response(
     *          response=403,
     *          description="Forbidden"
     *      )
     *)
     **/
    public function store(Gigquestion $request)
    {
        try{
            $input = $request->all();
            $input['userId'] = $request->user()->id;
            $gigquestion = $this->gigquestionInterface->create($input);
            return response()->json(['success' => true, 'gigquestion' => $gigquestion ]);
        }catch(Exception $ex)
        {
            return response()->json(['success' => false, 'error' => array('message' => $ex->getMessage() ) ]);
        }
    }

    /**
     * Display the specified resource.
     *
     * @param  int  $id
     * @return \Illuminate\Http\Response
     */
    public function show($id)
    {
        //
    }

    /**
     * Show the form for editing the specified resource.
     *
     * @param  int  $id
     * @return \Illuminate\Http\Response
     */
    public function edit($id)
    {
        //
    }

    /**
     * Update the specified resource in storage.
     *
     * @param  \Illuminate\Http\Request  $request
     * @param  int  $id
     * @return \Illuminate\Http\Response
     */

     /**
     * @OA\Put(
     ** path="/gigquestion/{questionId}",
     *   tags={"Gig Question"},
     *   summary="Edit Question",
     *   operationId="editquestion",
     *   security={{"bearer_token":{}}},
     *   @OA\Parameter(
     *          name="questionId",
     *          required=true,
     *          in="path",
     *          @OA\Schema(
     *              type="integer"
     *          )
     *   ),
     *
     *   @OA\RequestBody(
     *         required=true,
     *         @OA\MediaType(
     *             mediaType="application/json",
     *             @OA\Schema(
     *                 @OA\Property(
     *                     property="question",
     *                     type="string"
     *                 ),
     *                 @OA\Property(
     *                     property="gigId",
     *                     type="integer"
     *                 )
     *             )
     *         )
     *     ),
     *   @OA\Response(
     *      response=201,
     *       description="Success",
     *      @OA\MediaType(
     *           mediaType="application/json",
     *      )
     *   ),
     *   @OA\Response(
     *      response=400,
     *      description="Bad Request"
     *   ),
     *      @OA\Response(
     *          response=403,
     *          description="Forbidden"
     *      )
     *)
     **/
    public function update(Gigquestion $request, $id)
    {
        try{
            $exist = $this->gigquestionInterface->findOne(['id' => $id, 'userId' => auth()->user()->id ]);

            if(!$exist)
                return response()->json(['success' => false, 'error' => array('message' => "Data not Found" ) ]);

            $gigquestion = $this->gigquestionInterface->update($id, $request->all());
            return response()->json(['success' => true, 'gigquestion' => $gigquestion ]);
        }catch(Exception $ex)
        {
            return response()->json(['success' => false, 'error' => array('message' => $ex->getMessage() ) ]);

        }
    }

    /**
     * Remove the specified resource from storage.
     *
     * @param  int  $id
     * @return \Illuminate\Http\Response
     */

    /**
     * @OA\Delete(
     ** path="/gigquestion/{questionId}",
     *   tags={"Gig Question"},
     *   summary="Delete Question",
     *   operationId="deletequestion",
     *   security={{"bearer_token":{}}},
     *   @OA\Parameter(
     *          name="questionId",
     *          required=true,
     *          in="path",
     *          @OA\Schema(
     *              type="integer"
     *          )
     *   ),
     *   @OA\Response(
     *      response=201,
     *       description="Success",
     *      @OA\MediaType(
     *           mediaType="application/json",
     *      )
     *   ),
     *   @OA\Response(
     *      response=400,
     *      description="Bad Request"
     *   ),
     *      @OA\Response(
     *          response=403,
     *          description="Forbidden"
     *      )
     *)
     **/
    public function destroy($id)
    {
        try{
            $exist = $this->gigquestionInterface->findOne(['id' => $id, 'userId' => auth()->user()->id ]);

            if(!$exist)
                return response()->json(['success' => false, 'error' => array('message' => "Data not Found" ) ]);

            $this->gigquestionInterface->deleteById($id);
            return response()->json(['success' => true ]);
        }catch(Exception $ex)
        {
            return response()->json(['success' => false, 'error' => array('message' => $ex->getMessage() ) ]);

        }
    }
}